+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `options` property to `Advertisement`, containing `hideSponsoredBadge` field.
+
 ## [0.10.0] - 2023-09-27
 
 ## [0.9.0] - 2021-09-06